fix (irc): remove nick from kick message
Remove kicker nick from kick message when no other content is in the
message. Not sure if this is a user client thing, a network thing, or a
twisted irc thing, but if no message is specified in the kick, the
kicker nick was being used instead. This simply replaces it with '' if
it is the entire message.